fix: Move database-dependent Domain tests to Feature directory
- Moved all Domain service and activity tests that require database to Feature
- Updated namespaces from Tests\Unit to Tests\Feature
- Changed base class from UnitTestCase to TestCase for database tests
- Organized tests under proper Feature subdirectories matching Domain structure

This continues the effort to separate true unit tests from integration tests
that require database access.
