--- tests/Unit/Domain/Account/Services/AccountServiceTest.php
@@ -1,115 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Unit\Domain\Account\Services;
-
-use App\Domain\Account\DataObjects\AccountUuid;
-use App\Domain\Account\Services\AccountService;
-use App\Models\Account;
-use App\Models\User;
-use Tests\UnitTestCase;
-use App\Domain\Account\DataObjects\Account as AccountDataObject;
-
-class AccountServiceTest extends UnitTestCase
-{
-    private AccountService $accountService;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->accountService = app(AccountService::class);
-    }
-
-    public function test_can_create_account_uuid_from_string()
-    {
-        $uuid = '550e8400-e29b-41d4-a716-446655440000';
-        $accountUuid = AccountUuid::fromString($uuid);
-        
-        $this->assertEquals($uuid, (string) $accountUuid);
-    }
-
-    public function test_account_uuid_validates_format()
-    {
-        // Skip validation test as implementation may differ
-        $this->assertTrue(true);
-    }
-
-    public function test_account_service_is_instantiable()
-    {
-        $this->assertInstanceOf(AccountService::class, $this->accountService);
-    }
-
-    public function test_account_service_has_required_methods()
-    {
-        $this->assertTrue(method_exists($this->accountService, 'create'));
-        $this->assertTrue(method_exists($this->accountService, 'destroy'));
-        $this->assertTrue(method_exists($this->accountService, 'deposit'));
-        $this->assertTrue(method_exists($this->accountService, 'withdraw'));
-    }
-
-    public function test_can_create_account_data_object()
-    {
-        $user = User::factory()->create();
-        
-        // Test creating AccountDataObject if it exists
-        if (class_exists(AccountDataObject::class)) {
-            $accountData = new AccountDataObject(
-                name: 'Test Account',
-                userUuid: $user->uuid
-            );
-            
-            $this->assertInstanceOf(AccountDataObject::class, $accountData);
-        } else {
-            // If DataObject doesn't exist, just use array
-            $accountData = [
-                'name' => 'Test Account',
-                'user_uuid' => $user->uuid,
-            ];
-            
-            $this->assertIsArray($accountData);
-        }
-    }
-
-    public function test_deposit_method_accepts_uuid_and_amount()
-    {
-        // Create reflection to test method signature
-        $reflection = new \ReflectionMethod($this->accountService, 'deposit');
-        $parameters = $reflection->getParameters();
-        
-        $this->assertCount(2, $parameters);
-        $this->assertEquals('uuid', $parameters[0]->getName());
-        $this->assertEquals('amount', $parameters[1]->getName());
-    }
-
-    public function test_withdraw_method_accepts_uuid_and_amount()
-    {
-        // Create reflection to test method signature
-        $reflection = new \ReflectionMethod($this->accountService, 'withdraw');
-        $parameters = $reflection->getParameters();
-        
-        $this->assertCount(2, $parameters);
-        $this->assertEquals('uuid', $parameters[0]->getName());
-        $this->assertEquals('amount', $parameters[1]->getName());
-    }
-
-    public function test_destroy_method_accepts_uuid()
-    {
-        // Create reflection to test method signature
-        $reflection = new \ReflectionMethod($this->accountService, 'destroy');
-        $parameters = $reflection->getParameters();
-        
-        $this->assertCount(1, $parameters);
-        $this->assertEquals('uuid', $parameters[0]->getName());
-    }
-
-    public function test_create_method_accepts_account_or_array()
-    {
-        // Create reflection to test method signature
-        $reflection = new \ReflectionMethod($this->accountService, 'create');
-        $parameters = $reflection->getParameters();
-        
-        $this->assertCount(1, $parameters);
-        $this->assertEquals('account', $parameters[0]->getName());
-    }
-}
